refactor(likelihood): move base classes to root level module
Move core likelihood base classes from _base.py to
a new likelihood_base.py at the package root. This breaks a
cyclic dependency by decoupling the base classes from the likelihood
subpackage, allowing cleaner imports throughout the codebase.

All imports across the package have been updated to reference the new
module location.
