+++ firecrown/likelihood/factories/_builders.py
@@ -1,7 +1,7 @@
 """Factory methods for creating Likelihood objects from configuration."""
 
 from firecrown.likelihood.factories._models import TwoPointExperiment
-from firecrown.likelihood._base import Likelihood, NamedParameters
+from firecrown.likelihood_base import Likelihood, NamedParameters
 from firecrown.modeling_tools import ModelingTools
