+++ firecrown/likelihood/factories/_models.py
@@ -16,7 +16,7 @@ from firecrown.data_functions import (
     extract_all_real_data,
 )
 from firecrown.likelihood._gaussian import ConstGaussian
-from firecrown.likelihood._base import Likelihood
+from firecrown.likelihood_base import Likelihood
 from firecrown.likelihood._two_point import TwoPointFactory
 from firecrown.metadata_types import TwoPointCorrelationSpace
 from firecrown.utils import ensure_path, load_sacc_data
