+++ firecrown/likelihood/number_counts/_systematics.py
@@ -10,7 +10,7 @@ import pyccl
 
 from firecrown.updatable import register_new_updatable_parameter
 from firecrown.likelihood.number_counts._args import NumberCountsArgs
-from firecrown.likelihood._base import (
+from firecrown.likelihood_base import (
     SourceGalaxyPhotoZShift,
     SourceGalaxyPhotoZShiftandStretch,
     SourceGalaxySelectField,
